Fix CloudWatch scheduled event

+++ zappa/cli.py
@@ -20,8 +20,10 @@ import sys
 import tempfile
 import zipfile
 import pkg_resources
+import logging
+import botocore
 
-from zappa import Zappa
+from zappa import Zappa, logger
 
 CUSTOM_SETTINGS = [
     'aws_region',
@@ -35,13 +37,24 @@ CUSTOM_SETTINGS = [
     'touch',
 ]
 
+CLI_COMMANDS = [
+    'deploy',
+    'invoke',
+    'rollback',
+    'schedule',
+    'tail',
+    'undeploy',
+    'unschedule',
+    'update'
+]
+
 ##
 # Main Input Processing
 ##
 
 class ZappaCLI(object):
     """
-    ZappaCLI object is responsible for loading the settings, 
+    ZappaCLI object is responsible for loading the settings,
     handling the input arguments and executing the calls to the core library.
 
     """
@@ -64,6 +77,8 @@ class ZappaCLI(object):
     zip_path = None
     vpc_config = None
     memory_size = None
+    use_apigateway = None
+    lambda_handler = None
 
     def handle(self, argv=None):
         """
@@ -72,10 +87,10 @@ class ZappaCLI(object):
         Parses command, load settings and dispatches accordingly.
 
         """
+        help_message = "Please supply a command to execute. Can be one of: {}".format(', '.join(x for x in sorted(CLI_COMMANDS)))
 
         parser = argparse.ArgumentParser(description='Zappa - Deploy Python applications to AWS Lambda and API Gateway.\n')
-        parser.add_argument('command_env', metavar='U', type=str, nargs='*',
-                       help="Command to execute. Can be one of 'deploy', 'update', 'tail' and 'rollback'.")
+        parser.add_argument('command_env', metavar='U', type=str, nargs='*', help=help_message)
         parser.add_argument('-n', '--num-rollback', type=int, default=0,
                             help='The number of versions to rollback.')
         parser.add_argument('-s', '--settings_file', type=str, default='zappa_settings.json',
@@ -83,13 +98,15 @@ class ZappaCLI(object):
         parser.add_argument('-a', '--app_function', type=str, default=None,
                             help='The WSGI application function.')
         parser.add_argument('-v', '--version', action='store_true', help='Print the zappa version', default=False)
+        parser.add_argument('-y', action='store_true', help='Auto confirm yes', default=False)
 
         args = parser.parse_args(argv)
+
         vargs = vars(args)
         vargs_nosettings = vargs.copy()
         vargs_nosettings.pop('settings_file')
         if not any(vargs_nosettings.values()): # pragma: no cover
-            parser.error("Please supply a command to execute. Can be one of 'deploy', 'update', 'tail', rollback', 'invoke'.'")
+            parser.error(help_message)
             return
 
         # Version requires no arguments
@@ -100,6 +117,11 @@ class ZappaCLI(object):
         # Parse the input
         command_env = vargs['command_env']
         command = command_env[0]
+
+        if command not in CLI_COMMANDS:
+            print("The command '{}' is not recognized. {}".format(command, help_message))
+            return
+
         if len(command_env) < 2: # pragma: no cover
             self.load_settings_file(vargs['settings_file'])
 
@@ -126,21 +148,18 @@ class ZappaCLI(object):
         elif command == 'rollback': # pragma: no cover
             if vargs['num_rollback'] < 1: # pragma: no cover
                 parser.error("Please enter the number of iterations to rollback.")
-                return                
+                return
             self.rollback(vargs['num_rollback'])
         elif command == 'invoke': # pragma: no cover
             self.invoke()
         elif command == 'tail': # pragma: no cover
             self.tail()
         elif command == 'undeploy': # pragma: no cover
-            self.undeploy()
+            self.undeploy(noconfirm=args.y)
         elif command == 'schedule': # pragma: no cover
             self.schedule()
         elif command == 'unschedule': # pragma: no cover
             self.unschedule()
-        else:
-            print("The command '{0!s}' is not recognized.".format(command))
-            return
 
     ##
     # The Commands
@@ -175,7 +194,7 @@ class ZappaCLI(object):
         self.lambda_arn = self.zappa.create_lambda_function(bucket=self.s3_bucket_name,
                                                        s3_key=self.zip_path,
                                                        function_name=self.lambda_name,
-                                                       handler='handler.lambda_handler',
+                                                       handler=self.lambda_handler,
                                                        vpc_config=self.vpc_config,
                                                        timeout=self.timeout_seconds,
                                                        memory_size=self.memory_size)
@@ -184,19 +203,24 @@ class ZappaCLI(object):
         if self.zappa_settings[self.api_stage].get('keep_warm', True):
             self.zappa.create_keep_warm(self.lambda_arn, self.lambda_name)
 
-        # Create and configure the API Gateway
-        api_id = self.zappa.create_api_gateway_routes(
-            self.lambda_arn, self.lambda_name)
+        endpoint_url = ''
+        if self.use_apigateway:
+            # Create and configure the API Gateway
+            api_id = self.zappa.create_api_gateway_routes(
+                self.lambda_arn, self.lambda_name)
 
-        # Deploy the API!
-        cache_cluster_enabled = self.zappa_settings[self.api_stage].get('cache_cluster_enabled', False)
-        cache_cluster_size = str(self.zappa_settings[self.api_stage].get('cache_cluster_size', .5))
-        endpoint_url = self.zappa.deploy_api_gateway(
-                                    api_id=api_id, 
-                                    stage_name=self.api_stage,
-                                    cache_cluster_enabled=cache_cluster_enabled,
-                                    cache_cluster_size=cache_cluster_size
-                                )
+            # Deploy the API!
+            cache_cluster_enabled = self.zappa_settings[self.api_stage].get('cache_cluster_enabled', False)
+            cache_cluster_size = str(self.zappa_settings[self.api_stage].get('cache_cluster_size', .5))
+            endpoint_url = self.zappa.deploy_api_gateway(
+                                        api_id=api_id,
+                                        stage_name=self.api_stage,
+                                        cache_cluster_enabled=cache_cluster_enabled,
+                                        cache_cluster_size=cache_cluster_size
+                                    )
+
+            if self.zappa_settings[self.api_stage].get('touch', True):
+                requests.get(endpoint_url)
 
         # Finally, delete the local copy our zip package
         if self.zappa_settings[self.api_stage].get('delete_zip', True):
@@ -205,12 +229,8 @@ class ZappaCLI(object):
         # Remove the uploaded zip from S3, because it is now registered..
         self.zappa.remove_from_s3(self.zip_path, self.s3_bucket_name)
 
-        if self.zappa_settings[self.api_stage].get('touch', True):
-            requests.get(endpoint_url)
+        print("Deployed! {}".format(endpoint_url))
 
-        print("Your Zappa deployment is live!: " + endpoint_url)
-
-        return
 
     def update(self):
         """
@@ -326,20 +346,22 @@ class ZappaCLI(object):
         if self.zappa_settings[self.api_stage].get('events'):
             events = self.zappa_settings[self.api_stage]['events']
 
-            if type(events) != type([]):
+            if type(events) != list:
                 print("Events must be supplied as a list.")
                 return
 
-            # Update, as we need to get the Lambda ARN.
-            # There is probably a better way to do this.
-            # XXX
-            # http://boto3.readthedocs.io/en/latest/reference/services/lambda.html#Lambda.Client.get_function
-            self.update()
+            try:
+                function_response = self.zappa.lambda_client.get_function(FunctionName=self.lambda_name)
+            except botocore.exceptions.ClientError as e:
+                print("Function does not exist, please deploy first. Ex: zappa deploy {}".format(self.api_stage))
+                return
 
             print("Scheduling..")
-            self.zappa.schedule_events(self.lambda_arn, self.lambda_name, events)
+            self.zappa.schedule_events(
+                lambda_arn=function_response['Configuration']['FunctionArn'],
+                lambda_name=function_response['Configuration']['FunctionName'],
+                events=events)
 
-        return
 
     def unschedule(self):
         """
@@ -373,7 +395,7 @@ class ZappaCLI(object):
 
     def load_settings(self, settings_file="zappa_settings.json", session=None):
         """
-        Load the local zappa_settings.json file. 
+        Load the local zappa_settings.json file.
 
         An existing boto session can be supplied, though this is likely for testing purposes.
 
@@ -425,21 +447,17 @@ class ZappaCLI(object):
             self.api_stage].get('domain', None)
         self.timeout_seconds = self.zappa_settings[
             self.api_stage].get('timeout_seconds', 30)
+        self.use_apigateway = self.zappa_settings[
+            self.api_stage].get('use_apigateway', True)
+        self.lambda_handler = self.zappa_settings[
+            self.api_stage].get('lambda_handler', 'handler.lambda_handler')
 
-        # Create an Zappa object..
-        self.zappa = Zappa(session)
+        self.zappa = Zappa(boto_session=session, profile_name=self.profile_name, aws_region=self.aws_region)
 
-        # Explicitly set our AWS Region
-        self.zappa.aws_region = self.aws_region
-
-        # Load your AWS credentials from ~/.aws/credentials
-        self.zappa.load_credentials(session, self.profile_name)
-
-        # ..and configure it
         for setting in CUSTOM_SETTINGS:
             if setting in self.zappa_settings[self.api_stage]:
                 setattr(self.zappa, setting, self.zappa_settings[
-                        self.api_stage][setting])        
+                        self.api_stage][setting])
 
         return self.zappa
 
@@ -474,36 +492,35 @@ class ZappaCLI(object):
                 exclude=self.zappa_settings[self.api_stage].get('exclude', [])
             )
 
-        # Throw our setings into it
-        with zipfile.ZipFile(self.zip_path, 'a') as lambda_zip:
-            app_module, app_function = self.app_function.rsplit('.', 1)
-            settings_s = "# Generated by Zappa\nAPP_MODULE='{0!s}'\nAPP_FUNCTION='{1!s}'\n".format(app_module, app_function)
-            
-            if self.debug is not None:
-                settings_s = settings_s + "DEBUG='{0!s}'\n".format((self.debug)) # Cast to Bool in handler
-            settings_s = settings_s + "LOG_LEVEL='{0!s}'\n".format((self.log_level))
+        if self.app_function:
+            # Throw custom setings into the zip file
+            with zipfile.ZipFile(self.zip_path, 'a') as lambda_zip:
+                app_module, app_function = self.app_function.rsplit('.', 1)
+                settings_s = "# Generated by Zappa\nAPP_MODULE='{0!s}'\nAPP_FUNCTION='{1!s}'\n".format(app_module, app_function)
 
-            # If we're on a domain, we don't need to define the /<<env>> in
-            # the WSGI PATH
-            if self.domain:
-                settings_s = settings_s + "DOMAIN='{0!s}'\n".format((self.domain))
-            else:
-                settings_s = settings_s + "DOMAIN=None\n"
+                if self.debug:
+                    settings_s = settings_s + "DEBUG='{0!s}'\n".format((self.debug)) # Cast to Bool in handler
+                settings_s = settings_s + "LOG_LEVEL='{0!s}'\n".format((self.log_level))
 
-            # We can be environment-aware
-            settings_s = settings_s + "API_STAGE='{0!s}'\n".format((self.api_stage))
+                # If we're on a domain, we don't need to define the /<<env>> in
+                # the WSGI PATH
+                if self.domain:
+                    settings_s = settings_s + "DOMAIN='{0!s}'\n".format((self.domain))
+                else:
+                    settings_s = settings_s + "DOMAIN=None\n"
 
-            # Lambda requires a specific chmod
-            temp_settings = tempfile.NamedTemporaryFile(delete=False)
-            os.chmod(temp_settings.name, 0644)
-            temp_settings.write(settings_s)
-            temp_settings.close()
-            lambda_zip.write(temp_settings.name, 'zappa_settings.py')
-            os.remove(temp_settings.name)
+                # We can be environment-aware
+                settings_s = settings_s + "API_STAGE='{0!s}'\n".format((self.api_stage))
 
-            lambda_zip.close()
+                # Lambda requires a specific chmod
+                temp_settings = tempfile.NamedTemporaryFile(delete=False)
+                os.chmod(temp_settings.name, 0644)
+                temp_settings.write(settings_s)
+                temp_settings.close()
+                lambda_zip.write(temp_settings.name, 'zappa_settings.py')
+                os.remove(temp_settings.name)
+                # lambda_zip.close()
 
-        return
 
     def remove_local_zip(self):
         """
@@ -530,7 +547,7 @@ class ZappaCLI(object):
     def print_logs(self, logs):
         """
         Parse, filter and print logs to the console.
-        
+
         """
 
         for log in logs:
@@ -571,6 +588,7 @@ def handle(): # pragma: no cover
     """
     Main program execution handler.
     """
+
     try:
         cli = ZappaCLI()
         sys.exit(cli.handle())
